Fix reuse-after-free upon claim interface (libusb)
The libusb backend is too eager to report a device to the transport
layer. Because of the concept of "detached", this backend does not
try to open/claim interface. This can lead to a failure to Start()
upon registering the transport which in turns lead to use-after-free
when we send_connect after the transport is destroyed.

Bug: 372385779
Test: test_device.py

Change-Id: I424af66d1e1642d63723696bd228557db364cd56