Make bin2c_fastdeployment a cc_genrule
java_genrule is always built for device, but bin2c_fastdeployagent
was used by host cc modules. Changing it to a cc_genrule with
host_supported: true and enabling windows allows it to match its
variations with its reverse dependencies, which is important for
removing the 1-variant-fallback that this module was previously relying
on. Its dependency on deployagent still works because non-common arch modules can now depend on common arch modules.

Bug: 372091092
Test: Presubmits
Change-Id: Ied286b12b5e7408482518f353f27dcdb81416320