Fix Macos USB backend ZLP
ZLP was broken in aosp/2363263 but the revert was never merged.

Problem: We call GetEndpointProperties with a hard-coded kUSBOut
parameter. This call returns 0 for maxPacketSize in case of mismatch (
on a kUSBIn endpoint). Since endpoints on Pixels are returned in order
kUSBOut, then kUSBIn, we ALWAYS overwrite the correct value of
maxPacketSize (retrieve on the first endpoint) with 0.

This results in an incorrect mask of UINT_MAX and we NEVER send LZP. In
turn this leads to frozen USB comm when attempting to send a payload
multiple of the maxPacketSize (but less than reveiver RBO).

Fixes: 208675141
Fixes: 270205252
Test: No infra for it

Change-Id: Ice5a7a768d0bdf821d330e0e5676402ffc814ace